feat(Login+SignUp): Linked the two pages together so we can go from one to another.
We can now click on a "Register" link on the homepage to get redirected to the Sign Up page.
We can now click on a "Home" link on the homepage to get redirected to the home page.

Following a registration, an alert is now emitted on success or error (temporary feedback).